+++ src/test/exceptions_test.dart
@@ -7,6 +7,7 @@ import 'package:source_gen_test/source_gen_test.dart';
 void main() {
   LibraryReader library_1;
   LibraryReader library_2;
+  LibraryReader library_3;
   final path = testFilePath('test', 'src', 'model');
 
   setUp(() async {
@@ -14,6 +15,8 @@ void main() {
         path, "bad_nondbentity_probe.dart");
     library_2 = await initializeLibraryReaderForDirectory(
         path, "bad_no_datatable_annotation_probe.dart");
+    library_3 = await initializeLibraryReaderForDirectory(
+        path, "bad_multiple_datacolumns_on_field_test.dart");
   });
 
   group('Exceptions tests', () {
@@ -36,6 +39,16 @@ void main() {
       expect(
           v.contains(
               '''/*\nException: The DbEntity implementations must be annotated with `@DataTable`!'''),
+          true);
+    });
+
+    test('Test the multiple DataColumn annotations on the same field!',
+        () async {
+      String v = await generator.generate(library_3, null);
+
+      expect(
+          v.contains(
+              '''/*\nException while parsing field: description!\nException: Only one Column annotation is allowed on a Field!'''),
           true);
     });
   });