+++ src/CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 * Test for Account entity raw output generation
+* Test the multiple DataColumn annotations on the same field!
 
 ## [0.6.1] - 2019-05-06